apps/backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Autenticação
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Celery & Redis
celery==5.3.6
redis==5.0.1

# Azure SDK
azure-cognitiveservices-speech==1.34.1
openai==1.10.0
azure-identity==1.15.0

# OCI SDK
oci==2.119.1

# Vector Store
faiss-cpu==1.7.4
numpy==1.26.3

# Utils
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
tenacity==8.2.3

# Logging
loguru==0.7.2
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
ruff==0.1.13
mypy==1.8.0
black==23.12.1
